+++ app/src/main/java/com/ssu/bilda/presentation/home/TeamDetailsBySubjectFragment.kt
@@ -1,22 +1,20 @@
-package com.ssu.bilda
+package com.ssu.bilda.presentation.home
 
 import android.os.Bundle
-import androidx.fragment.app.Fragment
 import android.view.LayoutInflater
 import android.view.View
 import android.view.ViewGroup
-
-// TODO: Rename parameter arguments, choose names that match
-// the fragment initialization parameters, e.g. ARG_ITEM_NUMBER
-private const val ARG_PARAM1 = "param1"
-private const val ARG_PARAM2 = "param2"
+import androidx.fragment.app.Fragment
+import com.ssu.bilda.ARG_PARAM1
+import com.ssu.bilda.ARG_PARAM2
+import com.ssu.bilda.R
 
 /**
  * A simple [Fragment] subclass.
- * Use the [AddSubjectFragment.newInstance] factory method to
+ * Use the [TeamDetailsBySubjectFragment.newInstance] factory method to
  * create an instance of this fragment.
  */
-class AddSubjectFragment : Fragment() {
+class TeamDetailsBySubjectFragment : Fragment() {
     // TODO: Rename and change types of parameters
     private var param1: String? = null
     private var param2: String? = null
@@ -34,7 +32,7 @@ class AddSubjectFragment : Fragment() {
         savedInstanceState: Bundle?
     ): View? {
         // Inflate the layout for this fragment
-        return inflater.inflate(R.layout.fragment_add_subject, container, false)
+        return inflater.inflate(R.layout.fragment_team_details_by_subject, container, false)
     }
 
     companion object {
@@ -44,12 +42,12 @@ class AddSubjectFragment : Fragment() {
          *
          * @param param1 Parameter 1.
          * @param param2 Parameter 2.
-         * @return A new instance of fragment AddSubjectFragment.
+         * @return A new instance of fragment TeamDetailsBySubjectFragment.
          */
         // TODO: Rename and change types and number of parameters
         @JvmStatic
         fun newInstance(param1: String, param2: String) =
-            AddSubjectFragment().apply {
+            TeamDetailsBySubjectFragment().apply {
                 arguments = Bundle().apply {
                     putString(ARG_PARAM1, param1)
                     putString(ARG_PARAM2, param2)
