+++ app/src/main/res/layout/fragment_add_subject.xml
@@ -4,7 +4,7 @@
     xmlns:tools="http://schemas.android.com/tools"
     android:layout_width="match_parent"
     android:layout_height="match_parent"
-    tools:context=".AddSubjectFragment"
+    tools:context=".presentation.home.AddSubjectFragment"
     android:background="@color/bg_popup">
