+++ app/src/main/res/layout/fragment_team_details_by_subject.xml
@@ -4,7 +4,7 @@
     android:layout_width="match_parent"
     android:layout_height="match_parent"
     xmlns:app="http://schemas.android.com/apk/res-auto"
-    tools:context=".TeamDetailsBySubjectFragment">
+    tools:context=".presentation.home.TeamDetailsBySubjectFragment">
 
     <TextView
         android:id="@+id/tv_home_subject"
